seabee

Module constants

Constants

BPF_PATH

The path to the bpf filesystem

CONFIG_PATH

The default path to the configuration file

KEYLIST_PATH

The path where a list of keys is stored

KEY_DIR

The directory to store verification keys

KEY_SIGNATURE_DIR

The directory to store key signatures

PIN_DIR

The default directory name for pins in the BPF filesystem

POLICY_DIR

The directory to store policy files

POL_SIGNATURE_DIR

The directory to store policy signatures

SEABEECTL_EXE

The path to the seabeectl executable

SEABEE_DIR

The default directory for seabee files

SEABEE_MAX_MAPS

The max number of maps SeaBee can track concurrently. SeaBee tracks all maps on the system

SEABEE_MAX_POLICIES

The max number of policies seabee can store. Should be large enough for any use case.

SEABEE_MAX_POLICY_SCOPES

The max number of scopes for seabee policies.

SEABEE_ROOT_KEY_PATH

The location of the SeaBee root key, which is the first key to be loaded and is not verified with a signature.

SERVICE_PATH

SOCKET_PATH

SOCKET_TIMEOUT

SYS_PATH

The /sys filesystem