seabee::policy

Struct SeaBeePolicy

pub struct SeaBeePolicy {
    pub base_policy: PolicyFile,
    /* private fields */
}

Fields
base_policy: PolicyFile

SeaBee base policy (the policy that protects SeaBee itself)

Implementations

impl SeaBeePolicy

pub fn boot_time_key_load(&mut self) -> Result<()>

pub fn boot_time_policy_load(&mut self, maps: &SeaBeeMapHandles) -> Result<()>

overwrites all existing policies with policy from policy directory does not modify kernel structures

pub fn init_listener(&mut self) -> Result<()>

impl SeaBeePolicy

pub fn get_next_key_id(&mut self) -> Result<u32>

returns a new key id that does not yet exist

pub fn verify_new_policy( &self, new_policy: &mut PolicyFile, input_path: &PathBuf, sig_path: &Option<PathBuf>, ) -> Result<()>

pub fn verify_signature( &self, obj_path: &PathBuf, sig_path: &Option<PathBuf>, custom_digest: &Option<SeaBeeDigest>, key_id: Option<u32>, ) -> Result<()>

wrapper around crypto::verify_file_signature

This function will run even when verification is disabled, it will return Ok(()) and print a warning on verification failure

impl SeaBeePolicy

pub fn export_keys(&self) -> Result<()>

impl SeaBeePolicy

pub fn handle_cli_command( &mut self, cmd: SocketCommand, maps: &SeaBeeMapHandles, ) -> Result<String>

Execute a [PolicyCommand] from seabeectl

Returns and error or text that will be written back to the seabeectl client. Many different possible errors based on the command.

pub fn add_new_policy( &mut self, new_policy: &mut PolicyFile, pol_path: &PathBuf, sig_path: &Option<PathBuf>, maps: &SeaBeeMapHandles, ) -> Result<()>

add a new policy to SeaBee and to the kernel, but does not save to disk

pub fn remove_policy( &mut self, request: SignedRequestInfo, maps: &SeaBeeMapHandles, ) -> Result<()>

Removes a policy from SeaBee, the SeaBee policy folder, and from the kerenl

pub fn add_key( &mut self, src_path: &PathBuf, sig_path: &Option<PathBuf>, sig_digest: &Option<SeaBeeDigest>, ) -> Result<String>

pub fn get_key_by_path(&self, path: &PathBuf) -> Result<Option<&SeaBeeKey>>

O(n) operation that returns the corresponding SeaBeeKey if one exists or None

impl SeaBeePolicy

pub fn init(config: &Config) -> Result<Self>

pub fn get_unused_policy_id(&mut self) -> Result<u32>

returns a policy id that has not yet been used TODO: inflexible, will need to fit within kernel memory limits