seabee

Module seabeectl_lib

Structs

SeaBeeCtlArgs

Used to query or modify seaBee policies

SignedRequestInfo

Info needed to add, update, or remove a policy or key

Enums

CleanCommand

ConfigCommand

KeyIdentifier

LocalCommand

ObjIdentifier

Different ways to uniquely identify a SeaBee Object. Path method will only use the name field of the policy at the path as an identifier.

SeaBeeCtlCommands

SocketCommand

Functions

execute_args

execute_local_command

execute_socket_command

Executes the command passed to the command line that requires interaction with SeaBee via domain socket