seabee

Module utils

Statics

MACHINE_ID

Global OnceLock to hold the machine-id string.

Functions

create_dir_if_not_exists

ensure_root

Check if the process is running with uid 0

file_to_bytes

fill_buff_with_str

generate_sigmask

Generates a mask of allowed signals

init_machine_id

Initialize MACHINE_ID from /etc/machine-id if it isn’t already set.

open_or_create

try to open a file for reading, but create the file if it does not exist

remove_if_exists

str_to_abs_path_str

str_to_abs_pathbuf

str_to_bytes

Converts a path to a u8 vector with a requested max size

verify_file_has_ext

Error if file does not have extensions in list of expected_ext

verify_requirements

Ensure the system has all requirements for running ebpf security

verify_seabee_unloaded

Check that seabee programs have been unloaded

walk_with

Traverse all files and directories under root, calling f on each entry.