seabee

Module utils

Functions

create_dir_if_not_exists

ensure_root

Check if the process is running with uid 0

file_to_bytes

fill_buff_with_str

generate_sigmask

Generates a mask of allowed signals

is_sigint_allowed

open_or_create

try to open a file for reading, but create the file if it does not exist

remove_if_exists

str_to_abs_path

str_to_bytes

Converts a path to a u8 vector with a requested max size

verify_file_has_ext

Error if file does not have extensions in list of expected_ext

verify_requirements

Ensure the system has all requirements for running ebpf security